osupdater: clean up the files from removed boot dirs
imgbased cleans up /boot/ovirt-* when layers are removed, but any
files contained in that path which have been copied to /boot to
work around platform limitations with FIPS, virt-v2v, or kdump
are not, which can lead to /boot filling up on systems which have
had a large number of updates applied.

When we identify a path in /boot/ovirt-* to be removed, find any files
which were present in it, and remove them from /boot as well.

In addition, track down kdump kernels (which are dropped directly
in /boot) and remove those also.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1533931
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1533871
(cherry picked from commit 1a10a0ba16bac9e4dbc76e8cdf476edc02ad3d4b)

+++ src/imgbased/plugins/osupdater.py
@@ -910,7 +910,19 @@ def remove_boot(imgbase, lv_fullname):
     loader.remove_entry(lv_name)
 
     assert bootdir.strip("/") != "boot"
-    log.debug("Removing kernel dir: %s" % bootdir)
-    shutil.rmtree(bootdir)
+
+    bootfiles = [os.path.basename(b) for b in glob.glob("%s/*" % bootdir)]
+
+    bootfiles.extend([re.sub(r'(initramfs.*?).img', r'\1kdump.img', f)
+                      for f in bootfiles if "initramfs" in f])
+
+    for f in bootfiles:
+        if os.path.isfile("/boot/%s" % f):
+            log.debug("Removing extraneous boot file /boot/%s" % f)
+            os.unlink("/boot/%s" % f)
+
+    if os.path.exists(bootdir):
+        log.debug("Removing kernel dir: %s" % bootdir)
+        shutil.rmtree(bootdir)
 
 # vim: sw=4 et sts=4:
